+++ app/src/debug/res/values/google_maps_api.xml
@@ -1,0 +1,24 @@
+<resources>
+    <!--
+    TODO: Before you run your application, you need a Google Maps API key.
+
+    To get one, follow this link, follow the directions and press "Create" at the end:
+
+    https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=EB:99:95:BB:2A:C4:9F:E6:E4:18:6C:E5:2E:09:A9:23:8F:79:3C:2B%3Bcom.jjswigut.eventide.ui.map
+
+    You can also add your credentials to an existing key, using these values:
+
+    Package name:
+    com.jjswigut.eventide.ui.map
+
+    SHA-1 certificate fingerprint:
+    EB:99:95:BB:2A:C4:9F:E6:E4:18:6C:E5:2E:09:A9:23:8F:79:3C:2B
+
+    Alternatively, follow the directions here:
+    https://developers.google.com/maps/documentation/android/start#get-key
+
+    Once you have your key (it starts with "AIza"), replace the "google_maps_key"
+    string in this file.
+    -->
+    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">YOUR_KEY_HERE</string>
+</resources>
